EUI-4858-page-flow: refactor page flow name
