--- src/hearings/containers/request-hearing/additional-facilities/additional-facilities.component.html
@@ -1,3 +1,0 @@
-<p>
-  additional-facilities works!
-</p>
