--- src/hearings/containers/request-hearing/additional-instruction/additional-instruction.component.html
@@ -1,3 +1,0 @@
-<p>
-  additional-instruction works!
-</p>
