--- src/hearings/containers/request-hearing/check-answers/check-answers.component.html
@@ -1,3 +1,0 @@
-<p>
-  check-answers works!
-</p>
