+++ src/hearings/containers/request-hearing/hearing-additional-instructions/hearing-additional-instructions.component.html
@@ -1,0 +1,3 @@
+<p>
+  hearing-additional-instructions works!
+</p>
