+++ src/hearings/containers/request-hearing/hearing-attendance/hearing-attendance.component.html
@@ -1,0 +1,3 @@
+<p>
+  hearing-attendance works!
+</p>
