+++ src/hearings/containers/request-hearing/hearing-check-answers/hearing-check-answers.component.html
@@ -1,0 +1,3 @@
+<p>
+  hearing-check-answers works!
+</p>
