+++ src/hearings/containers/request-hearing/hearing-facilities/hearing-facilities.component.html
@@ -1,0 +1,3 @@
+<p>
+  hearing-facilities works!
+</p>
