+++ src/hearings/containers/request-hearing/hearing-facilities/hearing-facilities.component.ts
@@ -6,15 +6,15 @@ import * as fromHearingStore from '../../../store';
 import {RequestHearingPageFlow} from '../request-hearing.page.flow';
 
 @Component({
-  selector: 'exui-hearing-stage',
-  templateUrl: './hearing-stage.component.html',
+  selector: 'exui-hearing-facilities',
+  templateUrl: './hearing-facilities.component.html',
 })
-export class HearingStageComponent extends RequestHearingPageFlow implements OnInit, OnDestroy {
+export class HearingFacilitiesComponent extends RequestHearingPageFlow implements OnInit, OnDestroy {
 
   constructor(protected readonly hearingStore: Store<fromHearingStore.State>,
               protected readonly hearingsService: HearingsService) {
     super(hearingStore, hearingsService);
-  }á
+  }
 
   public ngOnInit(): void {
   }
@@ -34,3 +34,4 @@ export class HearingStageComponent extends RequestHearingPageFlow implements OnI
     super.unsubscribe();
   }
 }
+
