+++ src/hearings/containers/request-hearing/hearing-judge/hearing-judge.component.html
@@ -1,0 +1,3 @@
+<p>
+  hearing-judge works!
+</p>
