+++ src/hearings/containers/request-hearing/hearing-panel/hearing-panel.component.html
@@ -1,0 +1,3 @@
+<p>
+  hearing-panel works!
+</p>
