+++ src/hearings/containers/request-hearing/hearing-panel/hearing-panel.component.ts
@@ -6,15 +6,15 @@ import * as fromHearingStore from '../../../store';
 import {RequestHearingPageFlow} from '../request-hearing.page.flow';
 
 @Component({
-  selector: 'exui-hearing-stage',
-  templateUrl: './hearing-stage.component.html',
+  selector: 'exui-hearing-panel',
+  templateUrl: './hearing-panel.component.html',
 })
-export class HearingStageComponent extends RequestHearingPageFlow implements OnInit, OnDestroy {
+export class HearingPanelComponent extends RequestHearingPageFlow implements OnInit, OnDestroy {
 
   constructor(protected readonly hearingStore: Store<fromHearingStore.State>,
               protected readonly hearingsService: HearingsService) {
     super(hearingStore, hearingsService);
-  }á
+  }
 
   public ngOnInit(): void {
   }
