--- src/hearings/containers/request-hearing/hearing-requirement/hearing-requirement.component.html
@@ -1,3 +1,0 @@
-<p>
-  hearing-requirement works!
-</p>
