+++ src/hearings/containers/request-hearing/hearing-requirements/hearing-requirements.component.html
@@ -1,0 +1,3 @@
+<p>
+  hearing-requirements works!
+</p>
