+++ src/hearings/containers/request-hearing/hearing-venue/hearing-venue.component.scss
@@ -1,0 +1,3 @@
+.search-location {
+    max-width: 900px;
+}
