--- src/hearings/containers/request-hearing/how-party-attend/how-party-attend.component.html
@@ -1,3 +1,0 @@
-<p>
-  how-party-attend works!
-</p>
