+++ src/hearings/containers/request-hearing/index.ts
@@ -1,31 +1,29 @@
-import {AdditionalFacilitiesComponent} from './additional-facilities/additional-facilities.component';
-import {AdditionalInstructionComponent} from './additional-instruction/additional-instruction.component';
-import {CheckAnswersComponent} from './check-answers/check-answers.component';
-import {HearingRequirementComponent} from './hearing-requirement/hearing-requirement.component';
+import {HearingAdditionalInstructionsComponent} from './hearing-additional-instructions/hearing-additional-instructions.component';
+import {HearingAttendanceComponent} from './hearing-attendance/hearing-attendance.component';
+import {HearingCheckAnswersComponent} from './hearing-check-answers/hearing-check-answers.component';
+import {HearingFacilitiesComponent} from './hearing-facilities/hearing-facilities.component';
+import {HearingJudgeComponent} from './hearing-judge/hearing-judge.component';
+import {HearingPanelComponent} from './hearing-panel/hearing-panel.component';
+import {HearingRequirementsComponent} from './hearing-requirements/hearing-requirements.component';
 import {HearingStageComponent} from './hearing-stage/hearing-stage.component';
 import {HearingTimingComponent} from './hearing-timing/hearing-timing.component';
-import {HowPartyAttendComponent} from './how-party-attend/how-party-attend.component';
-import {LocationSearchComponent} from './location-search/location-search.component';
+import {HearingVenueComponent} from './hearing-venue/hearing-venue.component';
+import {HearingWelshComponent} from './hearing-welsh/hearing-welsh.component';
 import {RequestHearingComponent} from './request-hearing.component';
-import {RequirePanelOrNotComponent} from './require-panel-or-not/require-panel-or-not.component';
-import {SpecifyJudgeComponent} from './specify-judge/specify-judge.component';
-import {WelshHearingComponent} from './welsh-hearing/welsh-hearing.component';
 
 export const containers: any[] = [
   RequestHearingComponent,
-  AdditionalFacilitiesComponent,
-  WelshHearingComponent,
+  HearingFacilitiesComponent,
+  HearingWelshComponent,
   HearingTimingComponent,
-  HearingRequirementComponent,
+  HearingRequirementsComponent,
   HearingStageComponent,
-  LocationSearchComponent,
-  WelshHearingComponent,
-  HowPartyAttendComponent,
-  SpecifyJudgeComponent,
-  RequirePanelOrNotComponent,
+  HearingVenueComponent,
+  HearingWelshComponent,
+  HearingAttendanceComponent,
+  HearingJudgeComponent,
+  HearingPanelComponent,
   HearingTimingComponent,
-  AdditionalInstructionComponent,
-  CheckAnswersComponent
+  HearingAdditionalInstructionsComponent,
+  HearingCheckAnswersComponent
 ];
-
-
