--- src/hearings/containers/request-hearing/location-search/location-search.component.scss
@@ -1,3 +1,0 @@
-.search-location {
-    max-width: 900px;
-}
