--- src/hearings/containers/request-hearing/require-panel-or-not/require-panel-or-not.component.html
@@ -1,3 +1,0 @@
-<p>
-  require-panel-or-not works!
-</p>
