--- src/hearings/containers/request-hearing/specify-judge/specify-judge.component.html
@@ -1,3 +1,0 @@
-<p>
-  specify-judge works!
-</p>
