--- src/hearings/containers/request-hearing/specify-judge/specify-judge.component.spec.ts
@@ -1,51 +1,0 @@
-import {ComponentFixture, TestBed} from '@angular/core/testing';
-import {provideMockStore} from '@ngrx/store/testing';
-import {of} from 'rxjs';
-import {ACTION} from '../../../models/hearings.enum';
-import {HearingsService} from '../../../services/hearings.service';
-import {SpecifyJudgeComponent} from './specify-judge.component';
-
-describe('SpecifyJudgeComponent', () => {
-  let component: SpecifyJudgeComponent;
-  let fixture: ComponentFixture<SpecifyJudgeComponent>;
-  const mockedHttpClient = jasmine.createSpyObj('HttpClient', ['get', 'post']);
-  const hearingsService = new HearingsService(mockedHttpClient);
-  hearingsService.navigateAction$ = of(ACTION.CONTINUE);
-
-  const initialState = {
-    hearings: {
-      hearingList: {
-        caseHearingMainModel: [
-          {
-            hmctsServiceID: 'SSCS'
-          }
-        ]
-      },
-      hearingValues: null,
-      hearingRequest: null,
-      hearingConditions: null,
-    }
-  };
-
-  beforeEach(() => {
-    TestBed.configureTestingModule({
-      declarations: [SpecifyJudgeComponent],
-      providers: [
-        provideMockStore({initialState}),
-        {provide: HearingsService, useValue: hearingsService},
-      ]
-    }).compileComponents();
-
-    fixture = TestBed.createComponent(SpecifyJudgeComponent);
-    component = fixture.componentInstance;
-    fixture.detectChanges();
-  });
-
-  it('should create', () => {
-    expect(component).toBeTruthy();
-  });
-
-  afterEach(() => {
-    fixture.destroy();
-  });
-});
